requirements.txt
Flask==2.1.2
jpype1==1.2.1
gunicorn==20.1.0
python-Levenshtein==0.20.9
fuzzywuzzy==0.18.0
numpy==1.21.6
requests==2.27.1
gevent==21.12.0
gunicorn==20.1.0
python-dotenv==0.20.0
loguru==0.6.0
Flask-RESTful==0.3.9
Werkzeug==2.1.2
typing_extensions==4.2.0
minio==7.2.10